+++ src/sign_workflow/signers.py
@@ -6,16 +6,10 @@
 # this file be licensed under the Apache-2.0 license or a
 # compatible open source license.
 
-from typing import Any, List, Type, Union
 
+from sign_workflow.signer import Signer
 from sign_workflow.signer_pgp import SignerPGP
 from sign_workflow.signer_windows import SignerWindows
-from sign_workflow.signer import Signer
-
-"""
-This class is responsible for signing an artifact using the OpenSearch-signer-client and verifying its signature.
-The signed artifacts will be found in the same location as the original artifacts.
-"""
 
 
 class Signers:
